+++ internal/repository/poll_repo.go
@@ -20,7 +20,7 @@ type IPollRepo interface {
 	Create(params *CreatePollParams) (*model.Poll, error)
 	CreatePollOption(option *model.PollOption) error
 	FindByPollID(pollID string) (*model.Poll, error)
-	FindOpenedByMovieID(movieID int) (*model.Poll, error)
+	FindOpenedByMovieID(movieID int64) (*model.Poll, error)
 	FindPollOptionsByPollID(pollID int64) ([]*model.PollOption, error)
 	UpdateStatus(params *UpdateStatusParams) error
 }
@@ -52,7 +52,7 @@ func (r *PollRepo) FindByPollID(pollID string) (*model.Poll, error) {
 	return &poll, err
 }
 
-func (r *PollRepo) FindOpenedByMovieID(movieID int) (*model.Poll, error) {
+func (r *PollRepo) FindOpenedByMovieID(movieID int64) (*model.Poll, error) {
 	var poll model.Poll
 	err := r.db.Preload("Voting").Where("movie_id = ? AND status = ?", movieID, model.POLL_OPENED_STATUS).First(&poll).Error
 	if err != nil {
